Live Streaming Website with Admin Panel
Here, admin can upload news with image, videos and live video links so that it will appear on  client site. Client can only view the news page and will be redirected to the login page if they tried to view highlights or live  video if they are not logged in. After logging, client can interact with the content by comment, like or dislike on the video. Admin must create account for admin panel to control the entire system and interaction of the client.